Bugfix/delete exception

+++ src/main/java/com/mhealth/admin/service/HealthTipCategoryMasterService.java
@@ -7,8 +7,12 @@ import com.mhealth.admin.dto.request.HealthTipCategoryRequest;
 import com.mhealth.admin.dto.request.HealthTipCategorySearchRequest;
 import com.mhealth.admin.dto.response.PaginationResponse;
 import com.mhealth.admin.dto.response.Response;
+import com.mhealth.admin.model.HealthTip;
 import com.mhealth.admin.model.HealthTipCategoryMaster;
+import com.mhealth.admin.model.HealthTipPackageCategories;
 import com.mhealth.admin.repository.HealthTipCategoryMasterRepository;
+import com.mhealth.admin.repository.HealthTipPackageCategoriesRepository;
+import com.mhealth.admin.repository.HealthTipRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.data.domain.Page;
@@ -21,6 +25,7 @@ import org.springframework.util.ObjectUtils;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.Locale;
 import java.util.UUID;
 
@@ -32,6 +37,11 @@ public class HealthTipCategoryMasterService {
 
     @Autowired
     private MessageSource messageSource;
+
+    @Autowired
+    private HealthTipPackageCategoriesRepository healthTipPackageCategoriesRepository;
+    @Autowired
+    private HealthTipRepository healthTipRepository;
 
     public ResponseEntity<Response> addCategory(HealthTipCategoryRequest request, Locale locale) {
         String newFileName = uploadFile(request.getPhoto());
@@ -122,6 +132,22 @@ public class HealthTipCategoryMasterService {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
                     .body(new Response(Status.FAILED, Constants.NO_RECORD_FOUND_CODE,
                             messageSource.getMessage(Constants.HEALTH_TIP_CATEGORY_NOT_FOUND, null, locale)));
+        }
+        HealthTipCategoryMaster category = repository.findById(id).orElse(null);
+
+        // verify if category is used in any health tip package
+        HealthTipPackageCategories healthTipPackageCategories = healthTipPackageCategoriesRepository.findByHealthTipCategoryMaster(category);
+        if (healthTipPackageCategories != null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(new Response(Status.FAILED, Constants.FAILED_CODE,
+                            messageSource.getMessage(Constants.HEALTH_TIP_CATEGORY_USED_IN_HEALTH_TIP_PACKAGE, null, locale)));
+        }
+
+        List<HealthTip> healthTip = healthTipRepository.findByCategory(category);
+        if(!healthTip.isEmpty()){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(new Response(Status.FAILED, Constants.FAILED_CODE,
+                            messageSource.getMessage(Constants.HEALTH_TIP_CATEGORY_USED_IN_HEALTH_TIP_PACKAGE, null, locale)));
         }
 
         repository.deleteById(id);
